+++ ACorePoly/PolyLib/SegmentSolver.cpp
@@ -10,7 +10,7 @@ namespace ACorePolyLib
 	{
 		m_desc.desc = desc;
 		m_desc.cores.resize(params.param.size());
-		for (int i = 0; i < params.param.size(); ++i)
+		for (size_t i = 0; i < params.param.size(); ++i)
 		{
 			m_solvers.push_back( shared_ptr<ACoreSolver>( new ACoreSolver(m_desc.cores[i], params.param[i], desc) ) );
 		}
@@ -23,7 +23,7 @@ namespace ACorePolyLib
 	int SegmentSolver::GetDim() const
 	{
 		int res = 0;
-		for (int i = 0; i < m_solvers.size(); ++i)
+		for (size_t i = 0; i < m_solvers.size(); ++i)
 		{
 			res += m_solvers[i]->GetDim();
 		}
@@ -33,7 +33,7 @@ namespace ACorePolyLib
 
 	void SegmentSolver::FillBasis(int N, double * basis) const
 	{
-		for (int i = 0; i < m_solvers.size(); ++i)
+		for (size_t i = 0; i < m_solvers.size(); ++i)
 		{
 			m_solvers[i]->FillBasis(N, basis);
 			basis += N * m_solvers[i]->GetDim();
@@ -42,7 +42,7 @@ namespace ACorePolyLib
 
 	SegmentDescription & SegmentSolver::ReadResult(double * res)
 	{
-		for (int i = 0; i < m_solvers.size(); ++i)
+		for (size_t i = 0; i < m_solvers.size(); ++i)
 		{
 			m_solvers[i]->ReadResults(res);
 			res += m_solvers[i]->GetDim();
